activate proxy from config
activate proxy from config
