+++ src/server/index.tsx
@@ -5,12 +5,12 @@ import * as morgan from 'morgan';
 import * as promBundle from 'express-prom-bundle';
 import Helmet from 'react-helmet';
 import html from './html';
+import setUpDeltaProxy from './proxy';
 import silentRefreshHtml from './silent_refresh';
 import { RootState } from '../shared/store/reducers';
 import { DEFAULT_UI_SETTINGS } from '../shared/store/reducers/ui-settings';
 import { DEFAULT_SEARCH_CONFIG_PROJECT } from '../shared/store/reducers/config';
 import { DEFAULT_SEARCH_STATE } from '../shared/store/reducers/search';
-const { createProxyMiddleware } = require('http-proxy-middleware');
 
 const PORT_NUMBER = 8000;
 
@@ -57,9 +57,10 @@ app.use(`${base}/public`, express.static(join(__dirname, 'public')));
 // if in Dev mode, setup HMR and all the fancy stuff
 if (process.env.NODE_ENV !== 'production') {
   const { setupDevEnvironment } = require('./dev');
+  if (process.env.PROXY) {
+    setUpDeltaProxy(app, process.env.API_ENDPOINT || '');
+  }
   setupDevEnvironment(app);
-
-  // setUpProxyForSearch();
 }
 
 // silent refresh
@@ -81,7 +82,9 @@ app.get('*', async (req: express.Request, res: express.Response) => {
       pluginsManifestPath,
       subAppsManifestPath,
       dataModelsLocation,
-      apiEndpoint: process.env.API_ENDPOINT || '',
+      apiEndpoint: process.env.PROXY
+        ? '/proxy'
+        : process.env.API_ENDPOINT || '',
       basePath: base,
       clientId: process.env.CLIENT_ID || 'nexus-web',
       redirectHostName: `${process.env.HOST_NAME ||
@@ -112,268 +115,3 @@ app.listen(PORT_NUMBER, () => {
 });
 
 export default app;
-
-function setUpProxyForSearch() {
-  const esResult = {
-    hits: {
-      hits: [
-        {
-          _type: '_doc',
-          sort: [1627390182151],
-          _id:
-            'https://bbp.epfl.ch/neurosciencegraph/data/d52e34bd-5f62-495a-9e5c-b2d3afa33017',
-          _index: 'delta_3c07805d-0818-4eaf-8af4-91daa452268d_2',
-          _source: {
-            project: {
-              identifier: 'http://example.org/subjectSpecies',
-              label: 'Project',
-            },
-            '@type': ['a', 'b', 'c', 'd'],
-            name: 'A Resource',
-            description: 'Description',
-            subjectSpecies: {
-              identifier: 'http://example.org/subjectSpecies',
-              label: 'Subject Species',
-            },
-            brainRegion: {
-              identifier: 'http://example.org/subjectSpecies',
-              label: 'Brain Region',
-            },
-            contributors: [
-              {
-                identifier: 'http://example.org/subjectSpecies',
-                label: 'James Bond',
-              },
-              {
-                identifier: 'http://example.org/subjectSpecies',
-                label: 'Jason Bourne',
-              },
-            ],
-            organisations: [
-              {
-                identifier: 'http://example.org/subjectSpecies',
-                label: 'MI6',
-              },
-              {
-                identifier: 'http://example.org/subjectSpecies',
-                label: 'CIA',
-              },
-            ],
-            license: {
-              identifier: 'http://example.org/subjectSpecies',
-              label: 'License to Kill',
-            },
-            mType: [
-              {
-                identifier: 'http://example.org/subjectSpecies',
-                label: 'mType1',
-              },
-              {
-                identifier: 'http://example.org/subjectSpecies',
-                label: 'mType2',
-              },
-            ],
-          },
-        },
-      ],
-      max_score: 0.03460473,
-      total: {
-        relation: 'eq',
-        value: 115739,
-      },
-    },
-    timed_out: false,
-    took: 8,
-    _shards: {
-      failed: 0,
-      skipped: 0,
-      successful: 1,
-      total: 1,
-    },
-  };
-
-  type SearchConfig = {
-    fields: (
-      | {
-          name: string;
-          label: string;
-          array: boolean;
-          fields: { name: string; format: string }[];
-          format?: undefined;
-        }
-      | {
-          name: string;
-          label: string;
-          format: string;
-          array: boolean;
-          fields?: undefined;
-        }
-    )[];
-  };
-
-  const searchConfig: SearchConfig = {
-    fields: [
-      {
-        name: 'project',
-        label: 'Project',
-        array: false,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-      {
-        name: '@type',
-        label: 'Types',
-        format: 'uri',
-        array: true,
-      },
-      {
-        name: 'name',
-        label: 'Name',
-        array: false,
-        format: 'text',
-      },
-      {
-        name: 'description',
-        label: 'Description',
-        array: false,
-        format: 'text',
-      },
-      {
-        name: 'brainRegion',
-        label: 'Brain Region',
-        array: false,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-      {
-        name: 'subjectSpecies',
-        label: 'Subject Species',
-        array: false,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-      {
-        name: 'contributors',
-        label: 'Contributors',
-        array: true,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-      {
-        name: 'organisations',
-        label: 'Organisations',
-        array: true,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-      {
-        name: 'license',
-        label: 'License',
-        array: false,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-      {
-        name: 'mType',
-        label: 'M-Type',
-        array: true,
-        fields: [
-          {
-            name: 'identifier',
-            format: 'uri',
-          },
-          {
-            name: 'label',
-            format: 'text',
-          },
-        ],
-      },
-    ],
-  };
-
-  const filter = function(pathname: string, req: express.Request) {
-    console.log(`${req.method} ${pathname}`);
-    const query =
-      pathname.match('^/proxy/search/query') && req.method === 'POST';
-    const config =
-      pathname.match('^/proxy/search/config') && req.method === 'GET';
-    if (query || config) {
-      return false;
-    }
-    return true;
-  };
-
-  app.use(
-    '/proxy',
-    createProxyMiddleware(filter, {
-      target: process.env.API_ENDPOINT,
-      changeOrigin: true,
-      pathRewrite: {
-        [`^/proxy`]: '',
-      },
-    })
-  );
-  // search proxy
-  app.get(
-    `/proxy/search/config`,
-    (req: express.Request, res: express.Response) => {
-      res.send(JSON.stringify(searchConfig));
-    }
-  );
-
-  // search query
-  app.post(
-    `/proxy/search/query`,
-    (req: express.Request, res: express.Response) => {
-      res.send(JSON.stringify(esResult));
-    }
-  );
-}
